eval/generate_eval_data/requirements.txt
openai == 1.55.3
azure-search-documents == 11.6.0b2
python-dotenv == 1.0.1